+++ cpplib/graph/undirected_graph.hpp
@@ -4,7 +4,7 @@
 #include "dsu.hpp"
 #include "graph.hpp"
 
-template<typename T = long long, size_t MASK = 0>
+template<typename T = int64_t, size_t MASK = 0>
 class UndirectedGraph : public Graph<T, MASK> {
 public:
 	UndirectedGraph() : UndirectedGraph(0) {}
